+++ lib.esm/providers/abstract-provider.d.ts
@@ -60,7 +60,7 @@ export type DebugEventAbstractProvider = {
  *  if they are modifying a low-level feature of how subscriptions operate.
  */
 export type Subscription = {
-    type: "block" | "close" | "debug" | "error" | "network" | "pending";
+    type: "block" | "close" | "debug" | "error" | "finalized" | "network" | "pending" | "safe";
     tag: string;
 } | {
     type: "transaction";
@@ -208,6 +208,8 @@ export type PerformActionRequest = {
 } | {
     method: "getLogs";
     filter: PerformActionFilter;
+} | {
+    method: "getPriorityFee";
 } | {
     method: "getStorage";
     address: string;